requirements.txt
Django~=3.2.3
django-filter==2.4.0
djangorestframework==3.12.4
python-dotenv==0.17.1
django-phonenumber-field[phonenumbers]
djangorestframework
djangorestframework-camel-case==1.2.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt
